requirements.txt
asgiref==3.5.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.4
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.2
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.1.5
django-redis==5.2.0
django-six==1.0.4
djangorestframework==3.13.1
filelock==3.4.2
idna==3.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.4.1
protobuf==3.19.4
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dotenv==0.19.2
python-env==1.0.0
pytz==2021.3
redis==4.1.2
requests==2.27.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
wrapt==1.13.3
xlwt==1.3.0
